make `DependencyTest` resilient against line number changes
While it is generally nice to have classes as test input defined as local classes within the test method itself, this does not hold for tests that depend on line numbers. To avoid having to continuously change this test method for unrelated changes, just because the line numbers change, I have extracted the test class from the test method.
