+++ archunit/src/test/java/com/tngtech/archunit/core/domain/DependencyTest.java
@@ -9,6 +9,7 @@ import java.util.Set;
 
 import com.google.common.base.MoreObjects;
 import com.tngtech.archunit.base.DescribedPredicate;
+import com.tngtech.archunit.core.domain.testobjects.ClassWithArrayDependencies;
 import com.tngtech.archunit.core.domain.testobjects.ClassWithDependencyOnInstanceofCheck;
 import com.tngtech.archunit.core.domain.testobjects.ClassWithDependencyOnInstanceofCheck.InstanceOfCheckTarget;
 import com.tngtech.archunit.core.importer.ClassFileImporter;
@@ -86,26 +87,17 @@ public class DependencyTest {
     }
 
     @DataProvider
-    public static Object[][] method_calls_to_array_types() throws NoSuchMethodException {
-        @SuppressWarnings("unused")
-        class ClassWithArrayDependencies {
-            private void oneDimArray() {
-                new String[0].clone();
-            }
-
-            private void multiDimArray() {
-                new String[0][0].clone();
-            }
-        }
+    public static Object[][] method_calls_to_array_types() {
         return $$(
-                $(ClassWithArrayDependencies.class.getDeclaredMethod("oneDimArray"), String[].class, 93),
-                $(ClassWithArrayDependencies.class.getDeclaredMethod("multiDimArray"), String[][].class, 97)
+                $(ClassWithArrayDependencies.class, "oneDimArray", String[].class, 6),
+                $(ClassWithArrayDependencies.class, "multiDimArray", String[][].class, 10)
         );
     }
 
     @Test
     @UseDataProvider("method_calls_to_array_types")
-    public void Dependency_from_access_with_component_type(Method reflectionMethodWithArrayMethodCall, Class<?> arrayType, int expectedLineNumber) {
+    public void Dependency_from_access_with_component_type(Class<?> classDependingOnArray, String nameOfMethodWithArrayMethodCall, Class<?> arrayType, int expectedLineNumber) throws NoSuchMethodException {
+        Method reflectionMethodWithArrayMethodCall = classDependingOnArray.getDeclaredMethod(nameOfMethodWithArrayMethodCall);
         Class<?> reflectionDeclaringClass = reflectionMethodWithArrayMethodCall.getDeclaringClass();
         JavaMethod method = new ClassFileImporter().importClasses(reflectionDeclaringClass)
                 .get(reflectionDeclaringClass).getMethod(reflectionMethodWithArrayMethodCall.getName());
@@ -115,12 +107,12 @@ public class DependencyTest {
 
         DependenciesAssertion.ExpectedDependencies expectedDependencies = from(reflectionDeclaringClass).to(arrayType)
                 .withDescriptionContaining("Method <%s> calls method <%s>", method.getFullName(), arrayType.getName() + ".clone()")
-                .inLocation(DependencyTest.class, expectedLineNumber);
+                .inLocation(classDependingOnArray, expectedLineNumber);
         Class<?> expectedComponentType = arrayType.getComponentType();
         while (expectedComponentType != null) {
             expectedDependencies.from(reflectionDeclaringClass).to(expectedComponentType)
                     .withDescriptionContaining("Method <%s> depends on component type <%s>", method.getFullName(), expectedComponentType.getName())
-                    .inLocation(DependencyTest.class, expectedLineNumber);
+                    .inLocation(classDependingOnArray, expectedLineNumber);
             expectedComponentType = expectedComponentType.getComponentType();
         }
