+++ archunit/src/test/java/com/tngtech/archunit/core/domain/testobjects/ClassWithArrayDependencies.java
@@ -1,0 +1,12 @@
+package com.tngtech.archunit.core.domain.testobjects;
+
+@SuppressWarnings("unused")
+public class ClassWithArrayDependencies {
+    private void oneDimArray() {
+        new String[0].clone();
+    }
+
+    private void multiDimArray() {
+        new String[0][0].clone();
+    }
+}
